# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CadastraVision API
  description: "AI-enabled automated cadastral mapping platform. Programme: CadastralMap\
    \ \u2014 SIH 2026, PS 26012. Backend owner: Shiva."
  version: 0.2.0
paths:
...
  /v1/parcels/sync:
    post:
      tags:
      - Parcels
      summary: Offline sync batch
      description: "Submit a batch of offline actions. Each action requires a ``client_action_id``\
        \ for idempotent replay. Conflicts are surfaced explicitly \u2014 no silent\
        \ last-write-wins."
      operationId: sync_parcels_v1_parcels_sync_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SyncRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parcels:
    get:
      tags:
      - Parcels
      summary: List parcels
      description: 'Cursor-paginated list of parcels with optional filters. Filters:
        zone, jurisdiction, validation_status, confidence_min/max.'
      operationId: list_parcels_v1_parcels_get
      parameters:
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Pagination cursor
          title: Cursor
        description: Pagination cursor
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: zone
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Jurisdiction
      - name: validation_status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ParcelWorkflowStatus'
          - type: 'null'
          title: Validation Status
      - name: confidence_min
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Confidence Min
      - name: confidence_max
        in: query
        required: false
        schema:
          anyOf:
          - type: number
            maximum: 1.0
            minimum: 0.0
          - type: 'null'
          title: Confidence Max
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedResponse_ParcelSummary_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parcels/{parcel_id}:
    get:
      tags:
      - Parcels
      summary: Get parcel detail
      description: Retrieve full parcel representation including geometry.
      operationId: get_parcel_v1_parcels__parcel_id__get
      parameters:
      - name: parcel_id
        in: path
        required: true
        schema:
          type: string
          title: Parcel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParcelDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parcels/{parcel_id}/edit:
    post:
      tags:
      - Parcels
      summary: Edit parcel geometry
      description: Apply a vertex edit, split, or merge operation to a parcel. Triggers
        topology validation hook (Arun). ``idempotency_key`` prevents duplicate edits
        on retry.
      operationId: edit_parcel_v1_parcels__parcel_id__edit_post
      parameters:
      - name: parcel_id
        in: path
        required: true
        schema:
          type: string
          title: Parcel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParcelEditRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParcelEditResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parcels/{parcel_id}/approve:
    post:
      tags:
      - Parcels
      summary: Approve parcel
      description: "Human sign-off \u2014 marks the parcel as approved and eligible\
        \ for export. Requires the parcel to be in 'validated' status."
      operationId: approve_parcel_v1_parcels__parcel_id__approve_post
      parameters:
      - name: parcel_id
        in: path
        required: true
        schema:
          type: string
          title: Parcel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApproveRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParcelActionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parcels/{parcel_id}/reject:
    post:
      tags:
      - Parcels
      summary: Reject parcel
      description: Mark a parcel as rejected. Record is preserved for audit history.
      operationId: reject_parcel_v1_parcels__parcel_id__reject_post
      parameters:
      - name: parcel_id
        in: path
        required: true
        schema:
          type: string
          title: Parcel Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParcelActionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/validations/run:
    post:
      summary: Trigger validation run
      description: Trigger topology and confidence validation for a set of parcels.
      operationId: run_validation_v1_validations_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRunRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationRunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parcels/{parcel_id}/flags:
    get:
      tags:
      - Validation
      summary: Get validation flags for a parcel
      description: List all validation flags on a specific parcel.
      operationId: get_parcel_flags_v1_parcels__parcel_id__flags_get
      parameters:
      - name: parcel_id
        in: path
        required: true
        schema:
          type: string
          title: Parcel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParcelFlagsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/geo/parcels:
    get:
      tags:
      - Geographic Data
      summary: Get parcels as GeoJSON
      description: "Return all (or filtered) parcels as a GeoJSON FeatureCollection.\
        \ Supports bbox, zone, and jurisdiction filters. **STUB** \u2014 Phase 3:\
        \ real spatial query."
      operationId: get_geo_parcels_v1_geo_parcels_get
      parameters:
      - name: bbox
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Bounding box: minLon,minLat,maxLon,maxLat'
          title: Bbox
        description: 'Bounding box: minLon,minLat,maxLon,maxLat'
      - name: zone
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone
      - name: jurisdiction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Jurisdiction
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeoFeatureCollection'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ApproveRequest:
      properties:
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      title: ApproveRequest
      description: POST /v1/parcels/{parcel_id}/approve.
...
    EditOperation:
      properties:
        operation:
          type: string
          title: Operation
          description: vertex_edit | split | merge
        payload:
          additionalProperties: true
          type: object
          title: Payload
          description: Operation-specific geometry payload
      type: object
      required:
      - operation
      - payload
      title: EditOperation
      description: A single geometry edit operation.
...
    FlagResponse:
      properties:
        flag_id:
          type: string
          title: Flag Id
        parcel_id:
          type: string
          title: Parcel Id
        flag_type:
          $ref: '#/components/schemas/FlagType'
        severity:
          $ref: '#/components/schemas/FlagSeverity'
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        resolved:
          type: boolean
          title: Resolved
        created_at:
          type: string
          format: date-time
          title: Created At
        resolved_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Resolved At
      type: object
      required:
      - flag_id
      - parcel_id
      - flag_type
      - severity
      - resolved
      - created_at
      title: FlagResponse
      description: Single validation flag on a parcel.
    FlagSeverity:
      type: string
      enum:
      - info
      - warning
      - error
      - critical
      title: FlagSeverity
    FlagType:
      type: string
      enum:
      - overlap
      - gap
      - self_intersection
      - confidence_low
      - boundary_mismatch
      - other
      title: FlagType
    GeoFeature:
      properties:
        type:
          type: string
          title: Type
          default: Feature
        geometry:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Geometry
          description: GeoJSON geometry object (null if no geometry)
        properties:
          $ref: '#/components/schemas/GeoFeatureProperties'
      type: object
      required:
      - geometry
      - properties
      title: GeoFeature
    GeoFeatureCollection:
      properties:
        type:
          type: string
          title: Type
          default: FeatureCollection
        features:
          items:
            $ref: '#/components/schemas/GeoFeature'
          type: array
          title: Features
        total:
          type: integer
          title: Total
          default: 0
      type: object
      title: GeoFeatureCollection
      description: "GET /v1/geo/parcels \u2014 GeoJSON FeatureCollection."
      example:
        features: []
        total: 0
        type: FeatureCollection
    GeoFeatureProperties:
      properties:
        parcel_id:
          type: string
          title: Parcel Id
        workflow_status:
          type: string
          title: Workflow Status
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        zone:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone
      type: object
      required:
      - parcel_id
      - workflow_status
      title: GeoFeatureProperties
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PaginatedResponse_ParcelSummary_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ParcelSummary'
          type: array
          title: Items
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Opaque cursor token for the next page. Null when no more pages.
      type: object
      required:
      - items
      title: PaginatedResponse[ParcelSummary]
    ParcelActionResponse:
      properties:
        parcel_id:
          type: string
          title: Parcel Id
        workflow_status:
          $ref: '#/components/schemas/ParcelWorkflowStatus'
        message:
          type: string
          title: Message
        reviewed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Reviewed At
      type: object
      required:
      - parcel_id
      - workflow_status
      - message
      title: ParcelActionResponse
    ParcelDetail:
      properties:
        parcel_id:
          type: string
          title: Parcel Id
        geometry_wkt:
          anyOf:
          - type: string
          - type: 'null'
          title: Geometry Wkt
        source_tile_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Tile Id
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        zone:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone
        jurisdiction:
          anyOf:
          - type: string
          - type: 'null'
          title: Jurisdiction
        workflow_status:
          $ref: '#/components/schemas/ParcelWorkflowStatus'
        reviewed_by_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Reviewed By Id
        reviewed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Reviewed At
        review_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Notes
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - parcel_id
      - workflow_status
      - created_at
      - updated_at
      title: ParcelDetail
      description: Full parcel representation for GET /v1/parcels/{parcel_id}.
    ParcelEditRequest:
      properties:
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
          description: Client-generated key to make this edit idempotent
        edit:
          $ref: '#/components/schemas/EditOperation'
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - edit
      title: ParcelEditRequest
      description: POST /v1/parcels/{parcel_id}/edit.
    ParcelEditResponse:
      properties:
        parcel_id:
          type: string
          title: Parcel Id
        workflow_status:
          $ref: '#/components/schemas/ParcelWorkflowStatus'
        message:
          type: string
          title: Message
      type: object
      required:
      - parcel_id
      - workflow_status
      - message
      title: ParcelEditResponse
    ParcelFlagsResponse:
      properties:
        parcel_id:
          type: string
          title: Parcel Id
        flags:
          items:
            $ref: '#/components/schemas/FlagResponse'
          type: array
          title: Flags
        total:
          type: integer
          title: Total
      type: object
      required:
      - parcel_id
      - flags
      - total
      title: ParcelFlagsResponse
      description: GET /v1/parcels/{parcel_id}/flags.
    ParcelSummary:
      properties:
        parcel_id:
          type: string
          title: Parcel Id
        workflow_status:
          $ref: '#/components/schemas/ParcelWorkflowStatus'
        confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Confidence
        zone:
          anyOf:
          - type: string
          - type: 'null'
          title: Zone
        jurisdiction:
          anyOf:
          - type: string
          - type: 'null'
          title: Jurisdiction
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - parcel_id
      - workflow_status
      - created_at
      title: ParcelSummary
      description: Compact parcel representation for list endpoints.
    ParcelWorkflowStatus:
      type: string
      enum:
      - draft
      - validation_pending
      - validated
      - approved
      - rejected
      title: ParcelWorkflowStatus
...
    RejectRequest:
      properties:
        reason:
          type: string
          title: Reason
          description: Mandatory reason for rejection
      type: object
      required:
      - reason
      title: RejectRequest
      description: POST /v1/parcels/{parcel_id}/reject.
...
    SyncActionItem:
      properties:
        client_action_id:
          type: string
          title: Client Action Id
          description: Client-generated unique identifier for idempotency
        action_type:
          type: string
          title: Action Type
          description: edit | approve | reject | create
        parcel_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parcel Id
        payload:
          additionalProperties: true
          type: object
          title: Payload
      type: object
      required:
      - client_action_id
      - action_type
      title: SyncActionItem
      description: A single action in a sync batch.
    SyncRequest:
      properties:
        actions:
          items:
            $ref: '#/components/schemas/SyncActionItem'
          type: array
          maxItems: 100
          minItems: 1
          title: Actions
      type: object
      required:
      - actions
      title: SyncRequest
      description: "POST /v1/parcels/sync \u2014 offline action batch."
    SyncResponse:
      properties:
        applied:
          type: integer
          title: Applied
        conflicts:
          type: integer
          title: Conflicts
        failed:
          type: integer
          title: Failed
        results:
          items:
            $ref: '#/components/schemas/SyncResultItem'
          type: array
          title: Results
      type: object
      required:
      - applied
      - conflicts
      - failed
      - results
      title: SyncResponse
      description: Response from POST /v1/parcels/sync.
    SyncResultItem:
      properties:
        client_action_id:
          type: string
          title: Client Action Id
        status:
          type: string
          title: Status
          description: applied | conflict | failed
        parcel_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Parcel Id
        conflict_detail:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Conflict Detail
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      type: object
      required:
      - client_action_id
      - status
      title: SyncResultItem
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ValidationRunRequest:
      properties:
        parcel_ids:
          items:
            type: string
          type: array
          maxItems: 50
          minItems: 1
          title: Parcel Ids
          description: IDs of parcels to validate
      type: object
      required:
      - parcel_ids
      title: ValidationRunRequest
      description: POST /v1/validations/run.
    ValidationRunResponse:
      properties:
        job_id:
          type: string
          title: Job Id
        status:
          type: string
          title: Status
          default: queued
        parcel_count:
          type: integer
          title: Parcel Count
      type: object
      required:
      - job_id
      - parcel_count
      title: ValidationRunResponse